requirements.txt
pandas>=1.5.0
requests>=2.31.0
streamlit>=1.28.0
schedule>=1.2.0
plotly>=5.17.0
xgboost>=1.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
torch>=2.0.0
ta>=0.10.2
transformers>=4.30.0
vaderSentiment>=3.3.2
pytest>=7.4.0
